`_set_proofs_pending` performs DB related "proofs are spendable" check inside the lock.
