+++ cashu/mint/db/write.py
@@ -1,4 +1,3 @@
-import random
 from typing import List, Optional, Union
 
 from loguru import logger
@@ -32,42 +31,6 @@ class DbWriteHelper:
         self.crud = crud
         self.events = events
 
-    async def _set_proofs_pending(
-        self, proofs: List[Proof], quote_id: Optional[str] = None
-    ) -> None:
-        """If none of the proofs is in the pending table (_validate_proofs_pending), adds proofs to
-        the list of pending proofs or removes them. Used as a mutex for proofs.
-
-        Args:
-            proofs (List[Proof]): Proofs to add to pending table.
-            quote_id (Optional[str]): Melt quote ID. If it is not set, we assume the pending tokens to be from a swap.
-
-        Raises:
-            Exception: At least one proof already in pending table.
-        """
-        # first we check whether these proofs are pending already
-        random_id = random.randint(0, 1000000)
-        try:
-            logger.debug("trying to set proofs pending")
-            logger.trace(f"get_connection: random_id: {random_id}")
-            async with self.db.get_connection(
-                lock_table="proofs_pending",
-                lock_timeout=1,
-            ) as conn:
-                logger.trace(f"get_connection: got connection {random_id}")
-                await self._validate_proofs_pending(proofs, conn)
-                for p in proofs:
-                    logger.trace(f"crud: setting proof {p.Y} as PENDING")
-                    await self.crud.set_proof_pending(
-                        proof=p, db=self.db, quote_id=quote_id, conn=conn
-                    )
-                    logger.trace(f"crud: set proof {p.Y} as PENDING")
-        except Exception as e:
-            logger.error(f"Failed to set proofs pending: {e}")
-            raise TransactionError(f"Failed to set proofs pending: {str(e)}")
-        logger.trace("_set_proofs_pending released lock")
-        for p in proofs:
-            await self.events.submit(ProofState(Y=p.Y, state=ProofSpentState.pending))
 
     async def _unset_proofs_pending(self, proofs: List[Proof], spent=True) -> None:
         """Deletes proofs from pending table.
